requirements.txt
flask
pymupdf
langchain
faiss-cpu
sentence-transformers
tiktoken
python-dotenv
bert-score
openai
python-dotenv
